Fixes issue rancher/rancher#1923 * Added mixin for resizing dropdowns when dynamically adding content to the dropdown
Add some comments and rename a variable

refactor resize bootstrap dropdown function into util

make test pass
