+++ tests/unit/utils/bootstrap-fixes-test.js
@@ -1,0 +1,10 @@
+import bootstrapFixes from '../../../utils/bootstrap-fixes';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | bootstrap fixes');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = bootstrapFixes.resizeDropdown;
+  assert.ok(result);
+});
